+++ src/ui/dock.js
@@ -3,9 +3,9 @@
 import { store } from './store.js?v=e9304c5ee2';
 import { h, fmt, icon, svgIcon, popover, closePopover, clamp } from './util.js?v=13768f12bf';
 import { NODES, EDGES } from '../engine/topology.js?v=44e0aca402';
-import { createProfile, createScope, createSankey, createPerfusion } from './charts.js?v=52579fd006';
-import { createHVPG, createDoppler, createEndoscopy, createVarixWall, createAbdomen } from './instruments.js?v=d898e07e78';
-import { createLandscape } from './landscape.js?v=a9abd202bc';
+import { createProfile, createScope, createSankey, createPerfusion } from './charts.js?v=58ce81ca28';
+import { createHVPG, createDoppler, createEndoscopy, createVarixWall, createAbdomen } from './instruments.js?v=4f4af62718';
+import { createLandscape } from './landscape.js?v=8ae8f745bf';
 
 const NI = Object.fromEntries(NODES.map((n, i) => [n.id, i]));
 const EI = Object.fromEntries(EDGES.map((e, i) => [e.id, i]));